Re-ordered generated classes' members to follow guidelines (#220)
* Re-ordered generated classes' members to follow guidelines.

* Added support for nullable entity properties. These entities will not be created upon getting the property if null.

* Added test for nullable entity properties and fixed some failing tests

* Added description support for field instance and field pattern fields.

Integrated with SDK for retrieving those descriptions. Encompass will now be required to be installed on the machine in order to build the solution for the EntityGenerator project. SDK licensing is not required however.

Virtual field mappings are now generated directly in the EntityGenerator project.

* Small cleanup.

* Regenerated entities to get new entity classes and properties

* Generator no longer includes namespaces that aren't referenced.

* Fixed issue using BaseApiClient which wasn't occuring when using the developer connect endpoint.

Added new enum options.

Fixed some tests.