A BaseAddress URL can be set for API calls (#349)
* A BaseAddress URL can be set for API calls

ClientParameters now has a BaseAddress property that is used to set the BaseAddress value on the EncompassRestClient.  The ApiObject then uses the BaseAddress property to set its URL for API callouts.  The default value is "https://api.elliemae.com/".

* Changes as requested - BaseAddress not set to default value, and public setter added

EncompassRestClient now has a setter for BaseAddress.  This will allow for chaning the URL "after" the client instance is created.

Documentation added and modified as requested.

BaseAddress is now set to a default value in the EncompassRestClient constructor so as not to carry the value over during serialization.

* BaseAddress changed to non-nullable

Since the BaseAddress value is defaulted in the constructor and should never be null after that, changed to non-nullable.